requirements.txt
openai
python-dotenv
jinja2
